+++ src/module/actor/sheets/foesheet.ts
@@ -5,7 +5,6 @@ import { VueActorSheet } from '../../vue/vueactorsheet'
 export class FoeSheet extends VueActorSheet {
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
-      template: 'systems/foundry-ironsworn/templates/actor/foe.hbs',
       width: 450,
       height: 500,
     })
@@ -13,7 +12,7 @@ export class FoeSheet extends VueActorSheet {
 
   get renderHelperOptions(): Partial<VueSheetRenderHelperOptions> {
     return {
-      components: { 'foe-sheet': foeSheetVue },
+      rootComponent: foeSheetVue,
     }
   }