+++ src/module/item/move/sfmovesheet.ts
@@ -5,14 +5,13 @@ import { VueItemSheet } from '../../vue/vueitemsheet'
 export class SFMoveSheet extends VueItemSheet {
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
-      template: 'systems/foundry-ironsworn/templates/item/sfmove.hbs',
       height: 650,
     })
   }
 
   get renderHelperOptions(): Partial<VueSheetRenderHelperOptions> {
     return {
-      components: { 'sfmove-sheet': sfmoveSheetVue },
+      rootComponent: sfmoveSheetVue,
     }
   }
 }